tracee-rules: Remove duplicated code for testing Rego signatures
