--- tracee-rules/signatures/rego/examples/example1.rego
@@ -1,17 +1,0 @@
-package FOO_1
-
-__rego_metadoc__ := {
-	"id": "FOO-1",
-	"name": "example1"
-}
-
-tracee_selected_events[eventSelector] {
-	eventSelector := {
-		"source": "tracee",
-		#"name": "execve"
-	}
-}
-
-tracee_match {
-	endswith(input.args[0].value, "yo")
-}
